Update DefaultShader for GPU-side mesh data access, enhance AssetManager with detailed metadata handling, and improve EditorRenderer lighting setup.

+++ PegasusEditor/src/TabPanels/Viewport.cs
@@ -93,7 +93,7 @@ public class Viewport : TabPanel
             {
                 WindowSize = new(windowWidth, windowHeight);
                 
-                camera.AspectRatio = (float) windowWidth / windowHeight;
+                camera.AspectRatio = (float) windowWidth / (float) windowHeight;
                 
                 // TODO: Resize the framebuffer
             }
